Made the navigation menu its own component
