+++ v2/src/home/Welcome.tsx
@@ -1,48 +1,11 @@
 import UppercasedText from "../components/text/UppercasedText";
+import NavigationMenu from "./NavigationMenu";
 
 /**
  * Welcome and Navigation Component
  */
 function Welcome() {
 	const noLinkFields = ["Joseph Chen", "NYC, NY", "CS 3rd Year @ CWRU"];
-
-	const regLinkFields = [
-		{
-			title: "<About Me />",
-			href: "#aboutme",
-		},
-		{
-			title: "<Resume />",
-			href: "https://github.com/jchen42703/resumes/blob/main/latest_generic.pdf",
-		},
-		{
-			title: "<Contact Me />",
-			href: "#contact",
-		},
-	];
-
-	const workFields = [
-		{
-			title: "<Nemesis />",
-			href: "#nemesis",
-		},
-		{
-			title: "<Parallaxis (SWE) />",
-			href: "#parallaxis-swe",
-		},
-		{
-			title: "<Parallaxis (ML) />",
-			href: "#parallaxis-ml",
-		},
-		{
-			title: "<Machine Learning Research />",
-			href: "#machinelearning",
-		},
-		{
-			title: "<Applied ML (Hobby) />",
-			href: "#projects-mlops",
-		},
-	];
 
 	return (
 		<div className="grid grid-cols-4 content-container gap-x-0.75">
@@ -57,40 +20,7 @@ function Welcome() {
 				))}
 			</div>
 
-			<div className="col-span-1 flex flex-col">
-				<UppercasedText
-					text={"Menu"}
-					bold={true}
-					classes="text-lg"
-				></UppercasedText>
-				<br></br>
-				{regLinkFields.map(({ title, href }) => (
-					<UppercasedText
-						key={title}
-						text={title}
-						href={href}
-						bold={true}
-						classes="text-lg"
-					></UppercasedText>
-				))}
-				<br></br>
-
-				<UppercasedText
-					text={"Work & Research"}
-					bold={true}
-					classes="text-lg"
-				></UppercasedText>
-				<br></br>
-				{workFields.map(({ title, href }) => (
-					<UppercasedText
-						key={title}
-						text={title}
-						href={href}
-						bold={true}
-						classes="text-lg"
-					></UppercasedText>
-				))}
-			</div>
+			<NavigationMenu></NavigationMenu>
 
 			<div className="col-span-2 flex flex-col text-right">
 				{["Undergrad", "&", "(Ex) Founding SWE"].map((text) => (
